Refactor WrappedServer out of LlamaServer (#265)
* Refactor WrappedServer out of LlamaServer

* Add wrapped_server.py

* fixes

* Update src/lemonade/tools/server/wrapped_server.py


* suggestions from review

---------
